18.7. upload

+++ src/sk/tsystems/gamestudio/service/RatingService.java
@@ -5,7 +5,7 @@ import sk.tsystems.gamestudio.entity.Rating;
 public interface RatingService {
 	void add(Rating rating);
 
-	int getAverageRating(String game);
+	double getAverageRating(String game);
 	
 	int getNumberOfRatings(String game);
 }
